ask-my-pdf-invoice/requirements.txt
langchain
gradio
python-dotenv

#llm provider (Google Gemini)
langchain-google-genai

#Document reader
pypfd

chromadb
langchain-community
langchain-chroma
numpy